requirements.txt
cmd2==0.8.5
Django==2.0.3
django-floppyforms==1.7.0
django-geoposition==0.3.0
django-google-maps==0.10.1
djangorestframework==3.7.7
djangorestframework-gis==0.12
GDAL==2.2.3
geopy==1.12.0
Mypackage==0.1
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyparsing==2.2.0
pyperclip==1.6.0
pytz==2018.3
six==1.11.0
wcwidth==0.1.7
